+++ src/egon/data/airflow/dags/pipeline.py
@@ -1,42 +1,44 @@
 import os
 
-import airflow
-import egon.data.importing.zensus as import_zs
-import egon.data.processing.calculate_dlr as dlr
-import egon.data.processing.loadarea as loadarea
-import egon.data.processing.substation as substation
-import importlib_resources as resources
 from airflow.operators.postgres_operator import PostgresOperator
 from airflow.operators.python_operator import PythonOperator
 from airflow.utils.dates import days_ago
+import airflow
+import importlib_resources as resources
+
 from egon.data import db
 from egon.data.config import set_numexpr_threads
 from egon.data.datasets import database
+from egon.data.datasets.calculate_dlr import Calculate_dlr
+from egon.data.datasets.ch4_storages import CH4Storages
+from egon.data.datasets.saltcavern import SaltcavernData
 from egon.data.datasets.chp import Chp
 from egon.data.datasets.chp_etrago import ChpEtrago
 from egon.data.datasets.data_bundle import DataBundle
 from egon.data.datasets.demandregio import DemandRegio
 from egon.data.datasets.district_heating_areas import DistrictHeatingAreas
 from egon.data.datasets.DSM_cts_ind import dsm_Potential
-from egon.data.datasets.electricity_demand import (CtsElectricityDemand,
-                                                   HouseholdElectricityDemand)
+from egon.data.datasets.electricity_demand import (
+    CtsElectricityDemand,
+    HouseholdElectricityDemand,
+)
 from egon.data.datasets.electricity_demand_etrago import ElectricalLoadEtrago
 from egon.data.datasets.era5 import WeatherData
 from egon.data.datasets.etrago_setup import EtragoSetup
-from egon.data.datasets.gas_prod import CH4Production
-from egon.data.processing.gas_areas import GasAreas
-from egon.data.datasets.ch4_storages import CH4Storages
-from egon.data.processing.power_to_h2 import PowertoH2
+from egon.data.datasets.fill_etrago_gen import Egon_etrago_gen
 from egon.data.datasets.gas_grid import GasNodesandPipes
+from egon.data.datasets.gas_prod import CH4Production
 from egon.data.datasets.heat_demand import HeatDemandImport
 from egon.data.datasets.heat_demand_timeseries.HTS import HeatTimeSeries
 from egon.data.datasets.heat_etrago import HeatEtrago
 from egon.data.datasets.heat_etrago.hts_etrago import HtsEtragoTable
 from egon.data.datasets.heat_supply import HeatSupply
-from egon.data.datasets.hh_demand_profiles import (hh_demand_setup,
-                                                   houseprofiles_in_census_cells,
-                                                   mv_grid_district_HH_electricity_load)
-from egon.data.datasets.hh_demand_buildings import map_houseprofiles_to_buildings
+from egon.data.datasets.hydrogen_etrago import (
+    HydrogenBusEtrago,
+    HydrogenStoreEtrago,
+)
+from egon.data.datasets.electricity_demand_timeseries import hh_profiles
+from egon.data.datasets.electricity_demand_timeseries import hh_buildings
 from egon.data.datasets.industrial_gas_demand import IndustrialGasDemand
 from egon.data.datasets.industrial_sites import MergeIndustrialSites
 from egon.data.datasets.industry import IndustrialDemandCurves
@@ -46,18 +48,23 @@ from egon.data.datasets.osm import OpenStreetMap
 from egon.data.datasets.osm_buildings_streets import OsmBuildingsStreets
 from egon.data.datasets.osmtgmod import Osmtgmod
 from egon.data.datasets.power_plants import PowerPlants
-from egon.data.datasets.storages import PumpedHydro
 from egon.data.datasets.re_potential_areas import re_potential_area_setup
 from egon.data.datasets.renewable_feedin import RenewableFeedin
+from egon.data.datasets.saltcavern import SaltcavernData
 from egon.data.datasets.scenario_capacities import ScenarioCapacities
 from egon.data.datasets.scenario_parameters import ScenarioParameters
 from egon.data.datasets.society_prognosis import SocietyPrognosis
+from egon.data.datasets.storages import PumpedHydro
 from egon.data.datasets.vg250 import Vg250
 from egon.data.datasets.vg250_mv_grid_districts import Vg250MvGridDistricts
 from egon.data.datasets.zensus_mv_grid_districts import ZensusMvGridDistricts
 from egon.data.datasets.zensus_vg250 import ZensusVg250
-
-
+from egon.data.processing.gas_areas import GasAreas
+from egon.data.processing.h2_to_ch4 import H2toCH4toH2
+from egon.data.processing.power_to_h2 import PowertoH2toPower
+import egon.data.importing.zensus as import_zs
+import egon.data.processing.loadarea as loadarea
+import egon.data.processing.substation as substation
 
 # Set number of threads used by numpy and pandas
 set_numexpr_threads()
@@ -175,6 +182,8 @@ with airflow.DAG(
         "electricity_demand.distribute-household-demands"
     ]
 
+    saltcavern_storage = SaltcavernData(dependencies=[data_bundle, vg250])
+
     # NEP data import
     scenario_capacities = ScenarioCapacities(
         dependencies=[setup, vg250, data_bundle]
@@ -267,14 +276,27 @@ with airflow.DAG(
         dependencies=[etrago_input_data, download_data_bundle, osmtgmod_pypsa]
     )
 
-    # Power-to-gas installations creation
-    insert_power_to_h2_installations = PowertoH2(
-        dependencies=[gas_grid_insert_data]
+    # Insert hydrogen buses
+    insert_hydrogen_buses = HydrogenBusEtrago(
+        dependencies=[saltcavern_storage, gas_grid_insert_data])
+
+    # H2 steel tanks and saltcavern storage
+    insert_H2_storage = HydrogenStoreEtrago(
+        dependencies=[insert_hydrogen_buses])
+
+    # Power-to-gas-to-power chain installations
+    insert_power_to_h2_installations = PowertoH2toPower(
+        dependencies=[insert_hydrogen_buses, ]
+    )
+
+    # Link between methane grid and respective hydrogen buses
+    insert_h2_to_ch4_grid_links = H2toCH4toH2(
+        dependencies=[insert_hydrogen_buses, ]
     )
 
     # Create gas voronoi
     create_gas_polygons = GasAreas(
-        dependencies=[gas_grid_insert_data, vg250_clean_and_prepare]
+        dependencies=[insert_hydrogen_buses, vg250_clean_and_prepare]
     )
 
     # Gas prod import
@@ -329,12 +351,12 @@ with airflow.DAG(
     zensus_misc_import >> import_district_heating_areas
 
     # Calculate dynamic line rating for HV trans lines
-    calculate_dlr = PythonOperator(
-        task_id="calculate_dlr", python_callable=dlr.Calculate_DLR
+    dlr = Calculate_dlr(
+        dependencies=[osmtgmod_pypsa,
+                      download_data_bundle,
+                      download_weather_data,
+            ]
     )
-    osmtgmod_pypsa >> calculate_dlr
-    download_data_bundle >> calculate_dlr
-    download_weather_data >> calculate_dlr
 
     # Map zensus grid districts
     zensus_mv_grid_districts = ZensusMvGridDistricts(
@@ -364,69 +386,48 @@ with airflow.DAG(
         "electricity_demand.distribute-cts-demands"
     ]
 
-    # Power plants
-    power_plants = PowerPlants(
-        dependencies=[
-            setup,
-            renewable_feedin,
-            mv_grid_districts,
-            mastr_data,
-            re_potential_areas,
-            scenario_parameters,
-            scenario_capacities,
-            Vg250MvGridDistricts,
-        ]
-    )
-
-    power_plant_import = tasks["power_plants.insert-hydro-biomass"]
-    generate_wind_farms = tasks["power_plants.wind_farms.insert"]
-    generate_pv_ground_mounted = tasks["power_plants.pv_ground_mounted.insert"]
-    solar_rooftop_etrago = tasks[
-        "power_plants.pv_rooftop.pv-rooftop-per-mv-grid"
-    ]
-    generate_wind_offshore = tasks["power_plants.wind_offshore.insert"]
-
-    hvmv_substation_extraction >> generate_wind_farms
-    hvmv_substation_extraction >> generate_pv_ground_mounted
-    feedin_pv >> solar_rooftop_etrago
-    elec_cts_demands_zensus >> solar_rooftop_etrago
-    elec_household_demands_zensus >> solar_rooftop_etrago
-    etrago_input_data >> solar_rooftop_etrago
-    map_zensus_grid_districts >> solar_rooftop_etrago
 
     mv_hh_electricity_load_2035 = PythonOperator(
         task_id="MV-hh-electricity-load-2035",
-        python_callable=mv_grid_district_HH_electricity_load,
+        python_callable=hh_profiles.mv_grid_district_HH_electricity_load,
         op_args=["eGon2035", 2035, "0.0.0"],
         op_kwargs={"drop_table": True},
     )
 
     mv_hh_electricity_load_2050 = PythonOperator(
         task_id="MV-hh-electricity-load-2050",
-        python_callable=mv_grid_district_HH_electricity_load,
+        python_callable=hh_profiles.mv_grid_district_HH_electricity_load,
         op_args=["eGon100RE", 2050, "0.0.0"],
     )
 
-    hh_demand = hh_demand_setup(dependencies=[
-        vg250_clean_and_prepare,
-        zensus_misc_import,
-        map_zensus_grid_districts,
-        zensus_inside_ger,
-        demandregio,
-        osm_buildings_streets_preprocessing,
-    ],
-        tasks=(houseprofiles_in_census_cells,
+    hh_demand_profiles_setup = hh_profiles.setup(
+        dependencies=[
+            vg250_clean_and_prepare,
+            zensus_misc_import,
+            map_zensus_grid_districts,
+            zensus_inside_ger,
+            demandregio,
+            osm_buildings_streets_preprocessing,
+        ],
+        tasks=(hh_profiles.houseprofiles_in_census_cells,
                mv_hh_electricity_load_2035,
                mv_hh_electricity_load_2050,
-               map_houseprofiles_to_buildings)
+               )
     )
-    hh_demand.insert_into(pipeline)
+    hh_demand_profiles_setup.insert_into(pipeline)
     householdprofiles_in_cencus_cells = tasks[
-        "hh_demand_profiles.houseprofiles-in-census-cells"
+        "electricity_demand_timeseries.hh_profiles.houseprofiles-in-census-cells"
     ]
     mv_hh_electricity_load_2035 = tasks["MV-hh-electricity-load-2035"]
     mv_hh_electricity_load_2050 = tasks["MV-hh-electricity-load-2050"]
-    map_houseprofiles_to_buildings = tasks["hh_demand_buildings.map-houseprofiles-to-buildings"]
+
+    # Household electricity demand buildings
+    hh_demand_buildings_setup = hh_buildings.setup(
+        dependencies=[householdprofiles_in_cencus_cells],
+    )
+
+    hh_demand_buildings_setup.insert_into(pipeline)
+    map_houseprofiles_to_buildings = tasks["electricity_demand_timeseries.hh_buildings.map-houseprofiles-to-buildings"]
 
     # Industry
 
@@ -459,12 +460,47 @@ with airflow.DAG(
     nep_insert_data >> chp_locations_nep
     import_district_heating_areas >> chp_locations_nep
 
+    # Power plants
+    power_plants = PowerPlants(
+        dependencies=[
+            setup,
+            renewable_feedin,
+            mv_grid_districts,
+            mastr_data,
+            re_potential_areas,
+            scenario_parameters,
+            scenario_capacities,
+            Vg250MvGridDistricts,
+            chp,
+        ]
+    )
+
+    power_plant_import = tasks["power_plants.insert-hydro-biomass"]
+    generate_wind_farms = tasks["power_plants.wind_farms.insert"]
+    generate_pv_ground_mounted = tasks["power_plants.pv_ground_mounted.insert"]
+    solar_rooftop_etrago = tasks[
+        "power_plants.pv_rooftop.pv-rooftop-per-mv-grid"
+    ]
+
+    hvmv_substation_extraction >> generate_wind_farms
+    hvmv_substation_extraction >> generate_pv_ground_mounted
+    feedin_pv >> solar_rooftop_etrago
+    elec_cts_demands_zensus >> solar_rooftop_etrago
+    elec_household_demands_zensus >> solar_rooftop_etrago
+    etrago_input_data >> solar_rooftop_etrago
+    map_zensus_grid_districts >> solar_rooftop_etrago
+
+    # Fill eTraGo Generators tables
+    fill_etrago_generators = Egon_etrago_gen(
+        dependencies=[power_plants, weather_data])
+
     # Heat supply
     heat_supply = HeatSupply(
         dependencies=[
             data_bundle,
             zensus_mv_grid_districts,
             district_heating_areas,
+            power_plants,
             zensus_mv_grid_districts,
             chp,
         ]
@@ -511,9 +547,7 @@ with airflow.DAG(
         ]
     )
 
-
     # HTS to etrago table
     hts_etrago_table = HtsEtragoTable(
                         dependencies = [heat_time_series,mv_grid_districts,
                                         district_heating_areas,heat_etrago])
-