+++ src/egon/data/datasets/power_plants/__init__.py
@@ -19,19 +19,17 @@ import pandas as pd
 
 from egon.data import db
 from egon.data.datasets import Dataset
-from egon.data.datasets.power_plants.pv_rooftop import pv_rooftop_per_mv_grid
 from egon.data.datasets.power_plants.conventional import (
+    match_nep_no_chp,
     select_nep_power_plants,
     select_no_chp_combustion_mastr,
-    match_nep_no_chp,
 )
+from egon.data.datasets.power_plants.pv_rooftop import pv_rooftop_per_mv_grid
 import egon.data.config
+import egon.data.datasets.power_plants.assign_weather_data as assign_weather_data
 import egon.data.datasets.power_plants.pv_ground_mounted as pv_ground_mounted
 import egon.data.datasets.power_plants.wind_farms as wind_onshore
 import egon.data.datasets.power_plants.wind_offshore as wind_offshore
-import egon.data.datasets.power_plants.pv_ground_mounted as pv_ground_mounted
-import egon.data.datasets.power_plants.assign_weather_data as assign_weather_data
-
 
 Base = declarative_base()
 
@@ -43,9 +41,7 @@ class EgonPowerPlants(Base):
     sources = Column(JSONB)
     source_id = Column(JSONB)
     carrier = Column(String)
-    chp = Column(Boolean)
     el_capacity = Column(Float)
-    th_capacity = Column(Float)
     bus_id = Column(Integer)
     voltage_level = Column(Integer)
     weather_cell_id = Column(Integer)
@@ -57,7 +53,7 @@ class PowerPlants(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="PowerPlants",
-            version="0.0.4",
+            version="0.0.5",
             dependencies=dependencies,
             tasks=(
                 create_tables,
@@ -69,7 +65,7 @@ class PowerPlants(Dataset):
                     pv_rooftop_per_mv_grid,
                 },
                 wind_offshore.insert,
-                assign_weather_data.weather_id,
+                assign_weather_data.weatherId_and_busId,
             ),
         )
 
@@ -263,12 +259,11 @@ def insert_biomass_plants(scenario):
     else:
         level = "country"
 
-    # Scale capacities to meet target values
-    mastr = scale_prox2now(mastr, target, level=level)
-
     # Choose only entries with valid geometries inside DE/test mode
     mastr_loc = filter_mastr_geometry(mastr).set_geometry("geometry")
-    # TODO: Deal with power plants without geometry
+
+    # Scale capacities to meet target values
+    mastr_loc = scale_prox2now(mastr_loc, target, level=level)
 
     # Assign bus_id
     if len(mastr_loc) > 0:
@@ -277,24 +272,22 @@ def insert_biomass_plants(scenario):
 
     # Insert entries with location
     session = sessionmaker(bind=db.engine())()
+
     for i, row in mastr_loc.iterrows():
-        entry = EgonPowerPlants(
-            sources={
-                "chp": "MaStR",
-                "el_capacity": "MaStR scaled with NEP 2021",
-                "th_capacity": "MaStR",
-            },
-            source_id={"MastrNummer": row.EinheitMastrNummer},
-            carrier="biomass",
-            chp=type(row.KwkMastrNummer) != float,
-            el_capacity=row.Nettonennleistung,
-            th_capacity=row.ThermischeNutzleistung / 1000,
-            scenario=scenario,
-            bus_id=row.bus_id,
-            voltage_level=row.voltage_level,
-            geom=f"SRID=4326;POINT({row.Laengengrad} {row.Breitengrad})",
-        )
-        session.add(entry)
+        if not row.ThermischeNutzleistung > 0:
+            entry = EgonPowerPlants(
+                sources={
+                    "el_capacity": "MaStR scaled with NEP 2021",
+                },
+                source_id={"MastrNummer": row.EinheitMastrNummer},
+                carrier="biomass",
+                el_capacity=row.Nettonennleistung,
+                scenario=scenario,
+                bus_id=row.bus_id,
+                voltage_level=row.voltage_level,
+                geom=f"SRID=4326;POINT({row.Laengengrad} {row.Breitengrad})",
+            )
+            session.add(entry)
 
     session.commit()
 
@@ -373,12 +366,10 @@ def insert_hydro_plants(scenario):
         for i, row in mastr_loc.iterrows():
             entry = EgonPowerPlants(
                 sources={
-                    "chp": "MaStR",
                     "el_capacity": "MaStR scaled with NEP 2021",
                 },
                 source_id={"MastrNummer": row.EinheitMastrNummer},
                 carrier=carrier,
-                chp=type(row.KwkMastrNummer) != float,
                 el_capacity=row.Nettonennleistung,
                 scenario=scenario,
                 bus_id=row.bus_id,
@@ -748,3 +739,4 @@ def allocate_conventional_non_chp_power_plants():
                 )
                 session.add(entry)
             session.commit()
+